+++ changelog.md
@@ -4,6 +4,7 @@
 * Add `package.<package>.expose.target`, controlling whether a package is included in the default targets of envs.
 * Add the option `envs.<env>.localOverrides` to control whether local packages are included in an env's GHC overrides,
   independently of whether they're exposed.
+* Add an env, `hix-build-tools`, to configure packages used internally; like `cabal-install` and `hpack`.
 
 # 0.8.0